finalized generic apply to column util

+++ HPLC_MODULE_NOTEBOOK.py
@@ -3,6 +3,6 @@
 
 ######## GENERIC UTILITARIES ############
 def applyMappingToColumn(df, column, mapping):
-    df[column] = int(df[column])
+    df[column] = df[column].astype(int)
     df[column] = df.apply( lambda x: mapping[x[column]], axis = 1)
     return df
